#95 remove MediaStorage implementation
